requirements.txt
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.16.0
sqlparse==0.5.3


Django>=5.0.0
psycopg2-binary>=2.9.9  # For PostgreSQL
python-dotenv>=1.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-extensions>=3.2.0
whitenoise>=6.6.0